Moving writetojsonfile and adapting err msgs
