+++ pkg/util/util.go
@@ -4,9 +4,9 @@ import (
 	"archive/zip"
 	"bufio"
 	"bytes"
+	"encoding/json"
 	"errors"
 	"fmt"
-	devfilefs "github.com/devfile/library/pkg/testingutil/filesystem"
 	"hash/adler32"
 	"io"
 	"io/ioutil"
@@ -22,6 +22,8 @@ import (
 	"strings"
 	"syscall"
 	"time"
+
+	devfilefs "github.com/devfile/library/pkg/testingutil/filesystem"
 
 	"github.com/fatih/color"
 	"github.com/go-git/go-git/v5"
@@ -772,4 +774,22 @@ func IsPortFree(port int) bool {
 	_ = listener.Addr().(*net.TCPAddr).Port
 	err = listener.Close()
 	return err == nil
+}
+
+//WriteToJSONFile writes a struct to json file
+func WriteToJSONFile(c interface{}, filename string) error {
+	data, err := json.Marshal(c)
+	if err != nil {
+		return fmt.Errorf("unable to marshal data: %w", err)
+	}
+
+	if err = CreateIfNotExists(filename); err != nil {
+		return err
+	}
+	err = ioutil.WriteFile(filename, data, 0600)
+	if err != nil {
+		return fmt.Errorf("unable to write data to file %v: %w", c, err)
+	}
+
+	return nil
 }
